cloudrun_gcs_bq_gcs/requirements.txt
Flask
gunicorn
google-cloud-storage
google-cloud-bigquery